+++ backend/core/types/src/backend-swagger.ts
@@ -2535,6 +2535,9 @@ export interface Jurisdiction {
 
   /**  */
   emailFromAddress: string
+
+  /**  */
+  rentalAssistanceDefault: string
 }
 
 export interface AmiChart {
@@ -4435,6 +4438,9 @@ export interface JurisdictionCreate {
   emailFromAddress: string
 
   /**  */
+  rentalAssistanceDefault: string
+
+  /**  */
   programs: Id[]
 
   /**  */
@@ -4468,6 +4474,9 @@ export interface JurisdictionUpdate {
 
   /**  */
   emailFromAddress: string
+
+  /**  */
+  rentalAssistanceDefault: string
 
   /**  */
   programs: Id[]